backend/requirements.txt
twilio==8.10.0
phonenumbers==8.13.25
python-dotenv==1.0.0
Jinja2==3.1.2
